Move LabelSourceStateMeshPaneNode.populateGridWithMeshSettings into MeshSettingsNode
MeshSettingsNode is the only place where this method is called.
