+++ src/main/java/org/janelia/saalfeldlab/paintera/state/LabelSourceStateMeshPaneNode.kt
@@ -1,29 +1,27 @@
 package org.janelia.saalfeldlab.paintera.state
 
-import javafx.beans.property.BooleanProperty
-import javafx.beans.property.DoubleProperty
-import javafx.beans.property.IntegerProperty
-import javafx.beans.property.Property
-import javafx.beans.property.SimpleBooleanProperty
 import javafx.collections.FXCollections
 import javafx.collections.ListChangeListener
 import javafx.event.EventHandler
-import javafx.geometry.HPos
 import javafx.geometry.Pos
 import javafx.scene.Node
-import javafx.scene.control.*
-import javafx.scene.layout.*
-import javafx.scene.shape.CullFace
-import javafx.scene.shape.DrawMode
+import javafx.scene.control.Alert
+import javafx.scene.control.Button
+import javafx.scene.control.CheckBox
+import javafx.scene.control.Label
+import javafx.scene.control.TitledPane
+import javafx.scene.layout.HBox
+import javafx.scene.layout.Pane
+import javafx.scene.layout.Priority
+import javafx.scene.layout.VBox
 import javafx.stage.Modality
 import net.imglib2.type.label.LabelMultisetType
-import org.janelia.saalfeldlab.fx.Buttons
-import org.janelia.saalfeldlab.fx.Labels
 import org.janelia.saalfeldlab.fx.TitledPaneExtensions
-import org.janelia.saalfeldlab.fx.ui.NumericSliderWithField
 import org.janelia.saalfeldlab.fx.util.InvokeOnJavaFXApplicationThread
 import org.janelia.saalfeldlab.paintera.data.DataSource
-import org.janelia.saalfeldlab.paintera.meshes.*
+import org.janelia.saalfeldlab.paintera.meshes.GlobalMeshProgress
+import org.janelia.saalfeldlab.paintera.meshes.MeshInfo
+import org.janelia.saalfeldlab.paintera.meshes.MeshInfos
 import org.janelia.saalfeldlab.paintera.meshes.managed.MeshManagerWithAssignmentForSegments
 import org.janelia.saalfeldlab.paintera.meshes.managed.PainteraMeshManager
 import org.janelia.saalfeldlab.paintera.meshes.ui.MeshSettingsNode
@@ -34,10 +32,8 @@ import org.janelia.saalfeldlab.paintera.ui.source.mesh.MeshInfoNode
 import org.janelia.saalfeldlab.paintera.ui.source.mesh.MeshProgressBar
 import org.slf4j.LoggerFactory
 import java.lang.invoke.MethodHandles
-import java.util.*
+import java.util.Objects
 import java.util.stream.Collectors
-import kotlin.math.max
-import kotlin.math.min
 
 typealias TPE = TitledPaneExtensions
 
@@ -58,7 +54,7 @@ class LabelSourceStateMeshPaneNode(
             manager.rendererSettings.meshesEnabledProperty(),
             titledPaneGraphicsSettings = MeshSettingsNode.TitledPaneGraphicsSettings("Meshes"),
             helpDialogSettings = MeshSettingsNode.HelpDialogSettings(headerText = "Meshes"))
-        val contents = tp.content.asVBox()
+        tp.content.asVBox()
             .also { tp.content = it }
             .also { it.children.add(MeshesList(source, manager, meshInfos).node) }
         return tp
@@ -174,153 +170,6 @@ class LabelSourceStateMeshPaneNode(
     companion object {
 
         private val LOG = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass())
-
-		fun populateGridWithMeshSettings(
-				addMinLabelratioSlider: Boolean,
-				contents: GridPane,
-				initialRow: Int,
-                visibleCheckBox: CheckBox,
-				opacitySlider: NumericSliderWithField,
-				levelOfDetailSlider: NumericSliderWithField,
-				coarsestScaleLevelSlider: NumericSliderWithField,
-				finestScaleLevelSlider: NumericSliderWithField,
-				smoothingLambdaSlider: NumericSliderWithField,
-				smoothingIterationsSlider: NumericSliderWithField,
-				minLabelRatioSlider: NumericSliderWithField,
-				inflateSlider: NumericSliderWithField,
-				drawModeChoice: ComboBox<DrawMode>,
-				cullFaceChoice: ComboBox<CullFace>): Int {
-
-			setCoarsestAndFinestScaleLevelSliderListeners(
-					coarsestScaleLevelSlider.slider,
-					finestScaleLevelSlider.slider)
-
-			var row = initialRow
-
-			val textFieldWidth = 48.0
-			val choiceWidth = 95.0
-
-			// arrange the grid as 4 columns to fine-tune size and layout of the elements
-			for (i in 0..2)
-				contents.columnConstraints.add(ColumnConstraints())
-			contents.columnConstraints.add(ColumnConstraints(textFieldWidth))
-
-			val setupSlider = { slider: NumericSliderWithField, ttText: String ->
-				slider
-						.also { it.slider.isShowTickLabels = true }
-						.also { it.slider.tooltip = Tooltip(ttText) }
-						.also { it.textField.prefWidth = textFieldWidth }
-						.also { it.textField.maxWidth = Control.USE_PREF_SIZE }
-						.also { GridPane.setHgrow(it.slider, Priority.ALWAYS) }
-			}
-
-            contents.add(Labels.withTooltip("Visible"), 0, row)
-            contents.add(visibleCheckBox, 3, row)
-            GridPane.setHalignment(visibleCheckBox, HPos.CENTER)
-            ++row
-
-			contents.add(Labels.withTooltip("Opacity"), 0, row)
-			contents.add(opacitySlider.slider, 1, row)
-			GridPane.setColumnSpan(opacitySlider.slider, 2)
-			contents.add(opacitySlider.textField, 3, row)
-			setupSlider(opacitySlider, "Mesh Opacity")
-			++row
-
-			contents.add(Labels.withTooltip("Level of detail"), 0, row)
-			contents.add(levelOfDetailSlider.slider, 1, row)
-			GridPane.setColumnSpan(levelOfDetailSlider.slider, 2)
-			contents.add(levelOfDetailSlider.textField, 3, row)
-			setupSlider(levelOfDetailSlider, "Level Of Detail")
-			++row
-
-			contents.add(Labels.withTooltip("Coarsest scale"), 0, row)
-			contents.add(coarsestScaleLevelSlider.slider, 1, row)
-			GridPane.setColumnSpan(coarsestScaleLevelSlider.slider, 2)
-			contents.add(coarsestScaleLevelSlider.textField, 3, row)
-			setupSlider(coarsestScaleLevelSlider, "Coarsest Scale Level")
-			++row
-
-			contents.add(Labels.withTooltip("Finest scale"), 0, row)
-			contents.add(finestScaleLevelSlider.slider, 1, row)
-			GridPane.setColumnSpan(finestScaleLevelSlider.slider, 2)
-			contents.add(finestScaleLevelSlider.textField, 3, row)
-			setupSlider(finestScaleLevelSlider, "Finest Scale Level")
-			++row
-
-			contents.add(Labels.withTooltip("Lambda"), 0, row)
-			contents.add(smoothingLambdaSlider.slider, 1, row)
-			GridPane.setColumnSpan(smoothingLambdaSlider.slider, 2)
-			contents.add(smoothingLambdaSlider.textField, 3, row)
-			setupSlider(smoothingLambdaSlider, "Smoothing Lambda")
-			++row
-
-			contents.add(Labels.withTooltip("Iterations"), 0, row)
-			contents.add(smoothingIterationsSlider.slider, 1, row)
-			GridPane.setColumnSpan(smoothingIterationsSlider.slider, 2)
-			contents.add(smoothingIterationsSlider.textField, 3, row)
-			setupSlider(smoothingIterationsSlider, "Smoothing Iterations")
-			++row
-
-			// min label ratio slider only makes sense for sources of label multiset type
-			if (addMinLabelratioSlider)
-			{
-				contents.add(Labels.withTooltip("Min label ratio"), 0, row)
-				contents.add(minLabelRatioSlider.slider, 1, row)
-				GridPane.setColumnSpan(minLabelRatioSlider.slider, 2)
-				contents.add(minLabelRatioSlider.textField, 3, row)
-				setupSlider(minLabelRatioSlider, "Min label percentage for a pixel to be filled." + System.lineSeparator() +
-						"0.0 means that a pixel will always be filled if it contains the given label.")
-				++row
-			}
-
-			contents.add(Labels.withTooltip("Inflate"), 0, row)
-			contents.add(inflateSlider.slider, 1, row)
-			GridPane.setColumnSpan(inflateSlider.slider, 2)
-			contents.add(inflateSlider.textField, 3, row)
-			setupSlider(inflateSlider, "Inflate Meshes by Factor")
-			++row
-
-			val drawModeLabel = Labels.withTooltip("Draw Mode")
-			contents.add(drawModeLabel, 0, row)
-			GridPane.setColumnSpan(drawModeLabel, 2)
-			contents.add(drawModeChoice, 2, row)
-			GridPane.setColumnSpan(drawModeChoice, 2)
-			GridPane.setHalignment(drawModeChoice, HPos.RIGHT)
-			drawModeChoice.prefWidth = choiceWidth
-			++row
-
-			val cullFaceLabel = Labels.withTooltip("Cull Face")
-			contents.add(cullFaceLabel, 0, row)
-			GridPane.setColumnSpan(cullFaceLabel, 2)
-			contents.add(cullFaceChoice, 2, row)
-			GridPane.setColumnSpan(cullFaceChoice, 2)
-			GridPane.setHalignment(cullFaceChoice, HPos.RIGHT)
-			cullFaceChoice.prefWidth = choiceWidth
-			++row
-
-			return row
-		}
-
-		private fun setCoarsestAndFinestScaleLevelSliderListeners(
-				coarsestScaleLevelSlider: Slider,
-				finestScaleLevelSlider: Slider) {
-
-			coarsestScaleLevelSlider.valueProperty().addListener { _ ->
-				finestScaleLevelSlider.value = min(
-						coarsestScaleLevelSlider.value,
-						finestScaleLevelSlider.value)
-			}
-
-			finestScaleLevelSlider.valueProperty().addListener { _ ->
-				coarsestScaleLevelSlider.value = max(
-						coarsestScaleLevelSlider.value,
-						finestScaleLevelSlider.value)
-			}
-		}
-
-		private fun makeReloadSymbol() = RefreshButton
-				.createFontAwesome(scale = 2.0)
-				.also { it.rotate = 45.0 }
 
         private fun Node.asVBox() = if (this is VBox) this else VBox(this)
